rename: 컨벤션에 맞게 폴더 명 변경 기존 컴포넌트 폴더명 circular-icon-badges에서 카멜 케이스로 폴더명 변경
