+++ app/(anon)/test/style-badges/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import CircularIconBadge from '@/(anon)/_components/common/circular-icon-badges/CircularIconBadge';
+import CircularIconBadge from '@/(anon)/_components/common/circularIconBadges/CircularIconBadge';
 
 export default function StyleBadgesPage() {
   return (
